♻️ refactor(cli): centralize push command implementation
- move Push struct and its methods to cli/push.rs for better organization
- change run_push function to be a method of the Push struct for encapsulation

♻️ refactor(main): update push command execution

- modify main.rs to call the new run_push method on Push struct
